Split exec.go from run.go
